requirements.txt
boto3>=1.28.0
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
websockets>=12.0
python-multipart>=0.0.9
python-dotenv>=1.0.0
pyaudio>=0.2.14
aws-sdk-bedrock-runtime>=0.3.0
smithy-aws-core>=0.3.0
